feat: (WIP) add unit test github action to new fastapi
