+++ awast/cli/main.py
@@ -2,7 +2,8 @@ from enum import Enum
 from typing import List, Optional
 import typer
 from awast.cli.fastapi.new_api import new_api as new_fastapi
-from awast.cli.kubernetes.new_deployment import new_deployment, value_parser
+from awast.cli.kubernetes.new_deployment import new_deployment
+from awast.cli.utils import value_parser
 
 
 class Framework(str, Enum):
@@ -14,11 +15,15 @@ app = typer.Typer()
 
 
 @app.command(name="new-api")
-def new_api_cli(name: str, framework: Framework = Framework.fastapi):
+def new_api_cli(
+    name: str,
+    framework: Framework = Framework.fastapi,
+    set: List[str] = typer.Option([]),
+):
     print(f"Creating new API: {name}")
     print(f"Using templates for framework: {framework.value}")
     if framework.value == Framework.fastapi:
-        new_fastapi(name)
+        new_fastapi(name, set, value_parser)
     if framework.value == Framework.flask:
         raise NotImplementedError("Flask templates are not implemented yet")
     print("Success")
