+++ awast/templates/fastapi/app.py.jinja
