+++ pyproject.toml
@@ -21,6 +21,7 @@ pytest = "^7.2.0"
 requests = "^2.28.1"
 httpx = "^0.23.1"
 pytest-mock = "^3.10.0"
+coverage = "^6.5.0"
 
 [build-system]
 requires = ["poetry-core"]
